Refactored variable naming for clarity and improved file organization
