--- src/Controllers/includes/autoload.php
@@ -1,5 +1,0 @@
-<?php 
-function autoloader($className) {
-    require_once "class/$className.php";
-}
-spl_autoload_register("autoloader");
