+++ src/Controllers/loginFormHandler.php
@@ -18,7 +18,7 @@ if($_SERVER["REQUEST_METHOD"]=="POST"){
     }
     //other error handling can be added here
 
-    require_once("includes/config_session.inc.php");
+    require_once("includes/configSession.inc.php");
 
     if($errors){
         $_SESSION["login_errors"] = $errors;
